Implement a new style of input file parsing.
This new style of input parsing is a bit more suited to extracting
the needed information from the file and not trying to parse every
single byte of it.

In order to prove out the concept Advent of Code 2015 has been
updated to use it.  As part of that work a few minor changes to
methods used in other solutions have also been tweaked.